Removed library.cs and it cache settings
