+++ src/Umbraco.Tests/Configurations/UmbracoSettings/umbracoSettings.config
@@ -77,10 +77,6 @@
 
     <PreviewBadge><![CDATA[<a id="umbracoPreviewBadge" style="position: absolute; top: 0; right: 0; border: 0; width: 149px; height: 149px; background: url('{0}/assets/img/preview-mode-badge.png') no-repeat;" href="{0}/endPreview.aspx?redir={1}"><span style="display:none;">In Preview Mode - click to end</span></a>]]></PreviewBadge>
 
-    <!-- Cache cycle of Media and Member data fetched from the umbraco.library methods -->
-    <!-- In seconds. 0 will disable cache -->
-    <UmbracoLibraryCacheDuration>1800</UmbracoLibraryCacheDuration>
-
     <!-- How Umbraco should handle errors during macro execution. Can be one of the following values:
              - inline - show an inline error within the macro but allow the page to continue rendering. Historial Umbraco behaviour.
              - silent - Silently suppress the error and do not render the offending macro.
