+++ src/Umbraco.Tests/TestHelpers/SettingsForTests.cs
@@ -77,7 +77,6 @@ namespace Umbraco.Tests.TestHelpers
             settings.Setup(x => x.RequestHandler.AddTrailingSlash).Returns(true);
             settings.Setup(x => x.RequestHandler.UseDomainPrefixes).Returns(false);
             settings.Setup(x => x.RequestHandler.CharCollection).Returns(RequestHandlerElement.GetDefaultCharReplacements());
-            settings.Setup(x => x.Content.UmbracoLibraryCacheDuration).Returns(1800);
             settings.Setup(x => x.WebRouting.UrlProviderMode).Returns("AutoLegacy");
             settings.Setup(x => x.Templates.DefaultRenderingEngine).Returns(RenderingEngine.Mvc);
             settings.Setup(x => x.Providers.DefaultBackOfficeUserProvider).Returns("UsersMembershipProvider");
