+++ src/Umbraco.Web.UI/config/umbracoSettings.config
@@ -78,9 +78,6 @@
       <![CDATA[<a id="umbracoPreviewBadge" style="position: absolute; top: 0; right: 0; border: 0; width: 149px; height: 149px; background: url('{0}/assets/img/preview-mode-badge.png') no-repeat;" href="#" OnClick="javascript:window.top.location.href = '{0}/endPreview.aspx?redir={1}'"><span style="display:none;">In Preview Mode - click to end</span></a>
     ]]></PreviewBadge>
 
-    <!-- Cache cycle of Media and Member data fetched from the umbraco.library methods -->
-    <!-- In seconds. 0 will disable cache -->
-    <UmbracoLibraryCacheDuration>1800</UmbracoLibraryCacheDuration>
 
     <!-- Url Resolving ensures that all links works if you run Umbraco in virtual directories -->
     <!-- Setting this to true can increase render time for pages with a large number of links -->
